+++ src/CMakeLists.txt
@@ -1,1 +1,1 @@
-/opt/ros/kinetic/share/catkin/cmake/toplevel.cmake
+/opt/ros/melodic/share/catkin/cmake/toplevel.cmake
