--- src/tags:
@@ -1,12 +1,0 @@
-!_TAG_FILE_SORTED	2	/0=unsorted, 1=sorted, 2=foldcase/
-Imu	/Users/samflattery/dev/DuneBuggyROS/src/sensor_publisher.py	/^from sensor_msgs.msg import Imu, NavSatFix$/;"	i	language:Python
-NavSatFix	/Users/samflattery/dev/DuneBuggyROS/src/sensor_publisher.py	/^from sensor_msgs.msg import Imu, NavSatFix$/;"	i	language:Python
-parse_channel	/Users/samflattery/dev/DuneBuggyROS/src/sensor_publisher.py	/^def parse_channel(data_string, channel_number):$/;"	f	language:Python
-parse_GPS	/Users/samflattery/dev/DuneBuggyROS/src/sensor_publisher.py	/^def parse_GPS(data_string):$/;"	f	language:Python
-parse_IMU	/Users/samflattery/dev/DuneBuggyROS/src/sensor_publisher.py	/^def parse_IMU(data_string):$/;"	f	language:Python
-Quaternion	/Users/samflattery/dev/DuneBuggyROS/src/sensor_publisher.py	/^from geometry_msgs.msg import Quaternion$/;"	i	language:Python
-rospy	/Users/samflattery/dev/DuneBuggyROS/src/sensor_publisher.py	/^import rospy$/;"	i	language:Python
-sensor_publisher	/Users/samflattery/dev/DuneBuggyROS/src/sensor_publisher.py	/^def sensor_publisher(ser):$/;"	f	language:Python
-ser	/Users/samflattery/dev/DuneBuggyROS/src/sensor_publisher.py	/^    ser = serial.Serial()$/;"	v	language:Python
-serial	/Users/samflattery/dev/DuneBuggyROS/src/sensor_publisher.py	/^import serial$/;"	i	language:Python
-UInt8	/Users/samflattery/dev/DuneBuggyROS/src/sensor_publisher.py	/^from std_msgs.msg import UInt8$/;"	i	language:Python
